Initial commit V1.0 of SystemSage
This is the first version of SystemSage, a Python script designed to analyze and report on system hardware and software configurations. The script includes various functions to gather information about the system's CPU, RAM, storage, and network interfaces. It also generates a detailed report in both JSON and Markdown formats.
The script is intended for use in inventory management and system analysis, providing a comprehensive overview of the system's specifications.
The initial commit includes the following files:

The scope of the project is to provide a user-friendly tool for system administrators and IT professionals to quickly assess the hardware and software configurations of their systems. The script is designed to be easily extensible, allowing for future enhancements and additional features as needed.
I am going to add more features in the future, such as: The ability to remove orphaned files and registry keys, a desktop icon management system and features to move files and programs to a different drive.

I am no expert, this is a work in progress and just a hobby project to fill a personal want and need of my own. As I am trying to utilize open source software and libraries, I want to also create a project that is open source and free to use for anyone who wants to use it. I am not a programmer by trade, but I have been learning Python and programming in general for the past few years. I have created this project as a way to learn and practice my skills, and I hope that it can be useful to others as well.